+++ inc/customizer.php
@@ -10,7 +10,7 @@
  *
  * @param WP_Customize_Manager $wp_customize Theme Customizer object.
  */
-function cyclingclubcustomize_register( $wp_customize ) {
+function cyclingclub_customize_register( $wp_customize ) {
 		
 	$wp_customize->get_setting( 'blogname' )->transport         = 'postMessage';
 	$wp_customize->get_setting( 'blogdescription' )->transport  = 'postMessage';
@@ -29,7 +29,7 @@ function cyclingclubcustomize_register( $wp_customize ) {
 	
 	
 }
-add_action( 'customize_register', 'cyclingclubcustomize_register' );
+add_action( 'customize_register', 'cyclingclub_customize_register' );
 
 
 
@@ -38,7 +38,7 @@ add_action( 'customize_register', 'cyclingclubcustomize_register' );
  *
  * @return void
  */
-function cyclingclubcustomize_partial_blogname() {
+function cyclingclub_customize_partial_blogname() {
 	bloginfo( 'name' );
 }
 
@@ -47,14 +47,14 @@ function cyclingclubcustomize_partial_blogname() {
  *
  * @return void
  */
-function cyclingclubcustomize_partial_blogdescription() {
+function cyclingclub_customize_partial_blogdescription() {
 	bloginfo( 'description' );
 }
 
 /**
  * Binds JS handlers to make Theme Customizer preview reload changes asynchronously.
  */
-function cyclingclubcustomize_preview_js() {
+function cyclingclub_customize_preview_js() {
 	wp_enqueue_script( 'cyclingclub-customizer', get_template_directory_uri() . '/js/customizer.js', array( 'customize-preview' ), '20151215', true );
 }
-add_action( 'customize_preview_init', 'cyclingclubcustomize_preview_js' );
+add_action( 'customize_preview_init', 'cyclingclub_customize_preview_js' );