+++ inc/template-tags.php
@@ -7,11 +7,11 @@
  * @package cyclingclub
  */
 
-if ( ! function_exists( 'cyclingclubposted_on' ) ) :
+if ( ! function_exists( 'cyclingclub_posted_on' ) ) :
 	/**
 	 * Prints HTML with meta information for the current post-date/time.
 	 */
-	function cyclingclubposted_on() {
+	function cyclingclub_posted_on() {
 		$time_string = '<time class="entry-date published updated" datetime="%1$s">%2$s</time>';
 		if ( get_the_time( 'U' ) !== get_the_modified_time( 'U' ) ) {
 			$time_string = '<time class="entry-date published" datetime="%1$s">%2$s</time><time class="updated" datetime="%3$s">%4$s</time>';
@@ -35,11 +35,11 @@ if ( ! function_exists( 'cyclingclubposted_on' ) ) :
 	}
 endif;
 
-if ( ! function_exists( 'cyclingclubposted_by' ) ) :
+if ( ! function_exists( 'cyclingclub_posted_by' ) ) :
 	/**
 	 * Prints HTML with meta information for the current author.
 	 */
-	function cyclingclubposted_by() {
+	function cyclingclub_posted_by() {
 		$byline = sprintf(
 			/* translators: %s: post author. */
 			esc_html_x( 'by %s', 'post author', 'cyclingclub' ),
@@ -51,11 +51,11 @@ if ( ! function_exists( 'cyclingclubposted_by' ) ) :
 	}
 endif;
 
-if ( ! function_exists( 'cyclingclubentry_footer' ) ) :
+if ( ! function_exists( 'cyclingclub_entry_footer' ) ) :
 	/**
 	 * Prints HTML with meta information for the categories, tags and comments.
 	 */
-	function cyclingclubentry_footer() {
+	function cyclingclub_entry_footer() {
 		// Hide category and tag text for pages.
 		if ( 'post' === get_post_type() ) {
 			/* translators: used between list items, there is a space after the comma */
@@ -113,14 +113,14 @@ if ( ! function_exists( 'cyclingclubentry_footer' ) ) :
 	}
 endif;
 
-if ( ! function_exists( 'cyclingclubpost_thumbnail' ) ) :
+if ( ! function_exists( 'cyclingclub_post_thumbnail' ) ) :
 /**
  * Displays an optional post thumbnail.
  *
  * Wraps the post thumbnail in an anchor element on index views, or a div
  * element when on single views.
  */
-function cyclingclubpost_thumbnail() {
+function cyclingclub_post_thumbnail() {
 	if ( post_password_required() || is_attachment() || ! has_post_thumbnail() ) {
 		return;
 	}
@@ -135,13 +135,7 @@ function cyclingclubpost_thumbnail() {
 	<?php else : ?>
 
 	<a class="post-thumbnail" href="<?php the_permalink(); ?>" aria-hidden="true">
-		<?php
-			the_post_thumbnail( 'post-thumbnail', array(
-				'alt' => the_title_attribute( array(
-					'echo' => false,
-				) ),
-			) );
-		?>
+		<?php the_post_thumbnail( 'post-thumbnail' ); ?>
 	</a>
 
 	<?php endif; // End is_singular().
@@ -164,7 +158,7 @@ function cyclingclub_get_child_page_content() {
  *
  */
 
-function cyclingclubcomments($comment, $args, $depth) {
+function cyclingclub_comments($comment, $args, $depth) {
 
     if ( 'div' === $args['style'] ) {
         $tag       = 'div';