+++ template-parts/content-child-page.php
@@ -15,7 +15,7 @@
 		<?php the_title( '<h2 class="entry-title">', '</h2>' ); ?>
 	</header><!-- .entry-header -->
 
-	<?php cyclingclubpost_thumbnail(); ?>
+	<?php cyclingclub_post_thumbnail(); ?>
 
 	<div class="entry-content">